app/requirements.txt
bibtexparser==1.4.3
boto3==1.35.86
cuid2==2.0.1
Django==5.1.4
django-configurations==2.5.1
django-cors-headers==4.6.0
django-currentuser @ https://github.com/zsoldosp/django-currentuser/archive/refs/heads/master.zip
django-elasticsearch-dsl==7.4
django-filter==24.3
django-oauth-toolkit==3.0.1
django-maintenance-mode==0.21.1
django-recaptcha==4.0.0
django-sortedm2m==4.0.0
django-storages==1.14.4
django_structlog==9.0.0
django-tenants==3.7.0
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
drf-access-policy==1.5.0
drf-writable-nested==0.7.1
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
geopy==2.4.1
gunicorn==23.0.0
lxml==5.3.0
mysqlclient==2.1.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
pandas==2.2.3
passlib==1.7.4
psycopg==3.2.3
pycurl==7.45.4
pyhumps==3.8.0
pyzotero==1.5.28
queryish==0.2
requests==2.32.3
requests-aws4auth==1.3.1
stringcase==1.2.0
urllib3==1.26.18 # has to remain at this version because of elasticsearch
wagtail==6.4
wagtailcodeblock==1.29.0.2
wagtail-font-awesome-svg==1.0.1
Wand===0.6.13
